% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nums.R
\name{add_nums}
\alias{add_nums}
\title{Add Two Number}
\usage{
add_nums(x, y)
}
\arguments{
\item{x}{is a scalar to be added}

\item{y}{is a scalar to be added}
}
\value{
a scalar for the two numbers added together
}
\description{
This is a function that adds two scalar numbers
together
}
\examples{
x_num <- 4
y_num <- 9
add_nums(x_num, y_num)
}
