% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_data_year.R
\name{bind_data_year}
\alias{bind_data_year}
\title{process and bind multiple qualifying datasets}
\usage{
bind_data_year(..., years)
}
\arguments{
\item{...}{A list of data frames, each representing a qualifying lap
Each data frame should contain a \code{lap_number} column and its corresponding year.}

\item{years}{A numeric vector of years corresponding to each dataset in \code{...}.}
}
\value{
A single data frame with \code{lap_number} as a factor, a \code{year} column, and all laps combined.
}
\description{
process and bind multiple qualifying datasets
}
