% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor_color.R
\name{constructor_color}
\alias{constructor_color}
\title{Get the color associated with a constructor}
\usage{
constructor_color(constructor_input)
}
\arguments{
\item{constructor_input}{A character string representing either the full name of the constructor or a constructor abbreviation.}
}
\value{
A character string representing the color associated with the constructor. Returns NA if the constructor is not found.
}
\description{
Get the color associated with a constructor
}
