% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_utils.R
\name{corrected_track}
\alias{corrected_track}
\title{Adjust track plot aspect ratio}
\usage{
corrected_track(track_plot, track_coords)
}
\arguments{
\item{track_plot}{A ggplot2 object representing a track plot.}

\item{track_coords}{A data frame with columns x and y containing track coordinates.}
}
\value{
A ggplot2 plot object with corrected aspect ratio.
}
\description{
Adjust track plot aspect ratio
}
