% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_color.R
\name{driver_color}
\alias{driver_color}
\title{Get the color associated with a driver}
\usage{
driver_color(driver_input)
}
\arguments{
\item{driver_input}{A character string representing either the full name of the driver or a driver abbreviation.}
}
\value{
A character string representing the color associated with the driver. Returns NA if the driver is not found.
}
\description{
Get the color associated with a driver
}
