% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_utils.R
\name{smooth_track}
\alias{smooth_track}
\title{Generate a smooth track from lap details}
\usage{
smooth_track(lap_details)
}
\arguments{
\item{lap_details}{A data frame containing x and y coordinates of lap details. It should not include a 'time' column.}
}
\value{
A data frame with smoothed track coordinates.
}
\description{
Generate a smooth track from lap details
}
