% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_track.R
\name{theme_track}
\alias{theme_track}
\title{Custom ggplot2 Theme for Track Visualizations}
\usage{
theme_track()
}
\value{
A ggplot2 theme object customized for track visualizations.
}
\description{
Custom ggplot2 Theme for Track Visualizations
}
